refactor: add useTempoEditorKeyboardShortcut
