+++ app/src/components/TempoGraph/TempoEditor.tsx
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled"
 import { FC } from "react"
 import { TempoEditorScope } from "../../hooks/useTempoEditor"
-import { TempoEditorKeyboardShortcut } from "../KeyboardShortcut/TempoEditorKeyboardShortcut"
+import { useTempoEditorKeyboardShortcut } from "../../hooks/useTempoEditorKeyboardShortcut"
 import { TempoGraphToolbar } from "../TempoGraphToolbar/TempoGraphToolbar"
 import { TempoGraph } from "./TempoGraph"
 
@@ -10,13 +10,15 @@ const Container = styled.div`
   flex-direction: column;
   flex-grow: 1;
   overflow: hidden;
+  outline: none;
 `
 
 export const TempoEditor: FC = () => {
+  const keyboardShortcutProps = useTempoEditorKeyboardShortcut()
+
   return (
     <TempoEditorScope>
-      <Container>
-        <TempoEditorKeyboardShortcut />
+      <Container {...keyboardShortcutProps} tabIndex={0}>
         <TempoGraphToolbar />
         <TempoGraph />
       </Container>
